fix(app,shell): хвост красного CI 0510ec3 — needless_cast (win) + FSEvents-двойной Create (macOS)
Третий слой луковицы красного CI: после create_proxy-фикса clippy дошёл
до следующего линта в том же cfg(windows)-блоке T20 —
win32.hwnd.get() as isize (get() уже isize, raw-window-handle 0.6).
Каст снят; превентивно прогнана clippy-проба msvc-таргета для
canvas-widgets/core/render (чисто — это и есть T26-улучшение из плана
M7, применено раньше срока).

macOS: 8 из 9 вотчер-тестов вылечены фиксом 70783d1; остался
sync_dirs_idempotent — batch содержал 2 Create на одну запись (left: 2).
FSEvents может отдать два Create-флага (создание+закрытие) на одну
fs::write; строгий счёт ==1 — не инвариант продукта. Ассерт заменён на
инвариант множества путей: create_paths == {файл} + file_creates <= 2;
повторная регистрация вотча ловится напрямую (watched.len()==1).

Гейты: 599 тестов; fmt/clippy чистые; msvc clippy-проба чистая.

+++ crates/canvas-app/src/main.rs
@@ -4985,8 +4985,10 @@ impl ApplicationHandler<AppEvent> for App {
                             let user_data = canvas_shell::default_cache_dir()
                                 .unwrap_or_default()
                                 .join("webview2");
+                            // hwnd.get() уже isize (raw-window-handle 0.6):
+                            // без каста — иначе clippy needless_cast на Windows
                             self.widgets
-                                .attach_host(win32.hwnd.get() as isize, user_data, sender);
+                                .attach_host(win32.hwnd.get(), user_data, sender);
                         }
                     }
                     // На Windows бывает только Win32-handle
